+++ CHANGELOG.md
@@ -9,12 +9,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - New method `Toolchain::rustup_binary` to allow running arbitrary binaries managed by rustup. Before, only `rustc` and `cargo` could be run.
 
+### Changed
+
+- The default sandbox image is now fetched from [GitHub Container Registry][ghcr-linux].
+
+[ghcr-linux]: https://github.com/orgs/rust-lang/packages/container/package/crates-build-env/linux
+
 ## [0.12.0] - 2021-01-28
 
 ### Added
 
 - New variant `PrepareError::MissingDependencies`, returned during the prepare
   step when a dependency does not exist.
+
 ### Changed
 
 - Path dependencies are no longer removed from `Cargo.toml` during the prepare